make main activity detection more solid
