--- src/android/patches/0.17/addPackagePatch.js
@@ -1,2 +1,0 @@
-module.exports = (config) =>
-  `                .addPackage(${config.packageInstance})`;
