--- src/android/patches/0.18/addPackagePatch.js
@@ -1,2 +1,0 @@
-module.exports = (config) =>
-  ',\n        ' + config.packageInstance;
