+++ test/fixtures/android/0.17/patchedMainActivity.java
@@ -1,7 +1,6 @@
 package com.basic;
 
 import android.app.Activity;
-import com.oblador.vectoricons.VectorIconsPackage;
 import android.os.Bundle;
 import android.view.KeyEvent;
 
@@ -11,6 +10,8 @@ import com.facebook.react.ReactRootView;
 import com.facebook.react.modules.core.DefaultHardwareBackBtnHandler;
 import com.facebook.react.shell.MainReactPackage;
 import com.facebook.soloader.SoLoader;
+
+import com.oblador.vectoricons.VectorIconsPackage;
 
 public class MainActivity extends Activity implements DefaultHardwareBackBtnHandler {
